refactor: de-part region map shared palette and visibility (#4117)
Migrate palette tokens, visibility helpers, and label placement from
part libraries into explicit-import modules; shrink the no-part
grandfather by six entries while staying within the main-line budget.

Refs #4117
