+++ app/lib/features/game/flame/region_map/region_map_component_render_political_labels_province_compute.dart
@@ -28,7 +28,7 @@ extension _CtRegionMapRenderPoliticalLabelsProvinceCompute
           visibilityMode == CtMapVisibilityMode.playerConstrained,
     );
     if (rgb == null) {
-      return _provinceLabelPlateColor;
+      return RegionMapPalette.provinceLabelPlateColor;
     }
     return Color.fromRGBO(rgb.$1, rgb.$2, rgb.$3, kProvinceLabelPlateTintAlpha);
   }
