--- app/lib/features/game/flame/region_map/region_map_component_shared_visibility_labels.dart
@@ -1,62 +1,0 @@
-part of 'region_map_component.dart';
-
-/// Resolves map province-label icon ids.
-///
-/// Capital icon is prepended, then optional presence icons follow.
-List<String> resolveProvinceLabelIconIds({
-  required bool isCapital,
-  ProvinceUnitPresenceView? presence,
-}) {
-  final ids = <String>[if (isCapital) _provinceLabelCapitalIconId];
-  ids.addAll(resolveProvinceLabelPresenceIconIds(presence));
-  return ids;
-}
-
-/// Resolves optional prefix icon ids for sea-zone labels.
-List<String> resolveSeaZoneLabelPrefixIconIds({required bool isWarpZone}) {
-  if (!isWarpZone) {
-    return const <String>[];
-  }
-  return const <String>[_seaZoneLabelWarpIconId];
-}
-
-/// Resolves map province-label presence icon ids from province presence data.
-///
-/// Order is always civilian, regiment, ship. Icons are suppressed when intel is
-/// not visible or count is zero.
-List<String> resolveProvinceLabelPresenceIconIds(
-  ProvinceUnitPresenceView? presence,
-) {
-  final showPresenceIcons = presence != null && presence.intelVisible;
-  if (!showPresenceIcons) {
-    return const <String>[];
-  }
-  return <String>[
-    if (presence.civilianCount > 0) 'map_presence_civilian',
-    if (presence.regimentCount > 0) 'map_presence_regiment',
-    if (presence.shipCount > 0) 'map_presence_ship',
-  ];
-}
-
-/// Returns true when province-name text + icons should wrap icons to line 2.
-bool shouldWrapProvinceLabelPresenceIcons({
-  required double textWidthPx,
-  required int iconCount,
-  double maxWidthPx = _provinceLabelMaxWidthPx,
-  double iconRenderedPx = _provinceLabelIconRenderedPx,
-  double iconGapPx = _provinceLabelIconGapPx,
-  double textIconGapPx = _provinceLabelTextIconGapPx,
-}) {
-  if (iconCount <= 0) {
-    return false;
-  }
-  final iconsWidth =
-      (iconCount * iconRenderedPx) + ((iconCount - 1) * iconGapPx);
-  final singleLineContentWidth = textWidthPx + textIconGapPx + iconsWidth;
-  return singleLineContentWidth > maxWidthPx;
-}
-
-/// Capital labels must keep full text + star visible; do not ellipsize.
-bool shouldEllipsizeProvinceLabelText({required bool isCapital}) {
-  return !isCapital;
-}
