+++ tool/app_no_part_directives_grandfather.dart
@@ -83,12 +83,6 @@ const List<String> appNoPartDirectivesGrandfatheredForTests = <String>[
   'app/lib/features/game/flame/region_map/region_map_component_render_political_labels_province_paint.dart',
   'app/lib/features/game/flame/region_map/region_map_component_render_political_labels_sea.dart',
   'app/lib/features/game/flame/region_map/region_map_component_shared.dart',
-  'app/lib/features/game/flame/region_map/region_map_component_shared_label_placement.dart',
-  'app/lib/features/game/flame/region_map/region_map_component_shared_palette.dart',
-  'app/lib/features/game/flame/region_map/region_map_component_shared_visibility_extraction.dart',
-  'app/lib/features/game/flame/region_map/region_map_component_shared_visibility_fog_transport.dart',
-  'app/lib/features/game/flame/region_map/region_map_component_shared_visibility_halos.dart',
-  'app/lib/features/game/flame/region_map/region_map_component_shared_visibility_labels.dart',
   'app/lib/features/game/flame/region_map/region_map_province_overlay_geometry.dart',
   'app/lib/features/game/flame/region_map/region_map_province_overlay_geometry_segments.dart',
   'app/lib/features/game/flame/tilesets/terrain_tileset.dart',
